move requiredplugin to simulationcore (with its test + compat)
