+++ Component/SceneUtility/CMakeLists.txt
@@ -14,7 +14,6 @@ set(HEADER_FILES
     ${SOFACOMPONENTSCENEUTILITY_SOURCE_DIR}/MessageHandlerComponent.h
     ${SOFACOMPONENTSCENEUTILITY_SOURCE_DIR}/PauseAnimation.h
     ${SOFACOMPONENTSCENEUTILITY_SOURCE_DIR}/PauseAnimationOnEvent.h
-    ${SOFACOMPONENTSCENEUTILITY_SOURCE_DIR}/RequiredPlugin.h
 )
 
 set(SOURCE_FILES
@@ -27,7 +26,6 @@ set(SOURCE_FILES
     ${SOFACOMPONENTSCENEUTILITY_SOURCE_DIR}/MessageHandlerComponent.cpp
     ${SOFACOMPONENTSCENEUTILITY_SOURCE_DIR}/PauseAnimation.cpp
     ${SOFACOMPONENTSCENEUTILITY_SOURCE_DIR}/PauseAnimationOnEvent.cpp
-    ${SOFACOMPONENTSCENEUTILITY_SOURCE_DIR}/RequiredPlugin.cpp
 )
 
 sofa_find_package(SofaSimulationCore REQUIRED)
