+++ SofaKernel/modules/SofaBaseUtils/SofaBaseUtils_test/CMakeLists.txt
@@ -6,7 +6,6 @@ set(SOURCE_FILES
         MessageHandlerComponent_test.cpp
         MakeAliasComponent_test.cpp
         MakeDataAliasComponent_test.cpp
-        RequiredPlugin_test.cpp
         AddResourceRepository_test.cpp
 )
