+++ SofaKernel/modules/SofaSimulationCore/SofaSimulationCore_test/RequiredPlugin_test.cpp
@@ -1,5 +1,26 @@
-#include <sofa/testing/BaseTest.h>
-using sofa::testing::BaseTest;
+/******************************************************************************
+*                 SOFA, Simulation Open-Framework Architecture                *
+*                    (c) 2006 INRIA, USTL, UJF, CNRS, MGH                     *
+*                                                                             *
+* This program is free software; you can redistribute it and/or modify it     *
+* under the terms of the GNU Lesser General Public License as published by    *
+* the Free Software Foundation; either version 2.1 of the License, or (at     *
+* your option) any later version.                                             *
+*                                                                             *
+* This program is distributed in the hope that it will be useful, but WITHOUT *
+* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or       *
+* FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License *
+* for more details.                                                           *
+*                                                                             *
+* You should have received a copy of the GNU Lesser General Public License    *
+* along with this program. If not, see <http://www.gnu.org/licenses/>.        *
+*******************************************************************************
+* Authors: The SOFA Team and external contributors (see Authors.txt)          *
+*                                                                             *
+* Contact information: contact@sofa-framework.org                             *
+******************************************************************************/
+#include <sofa/testing/BaseSimulationTest.h>
+using sofa::testing::BaseSimulationTest;
 
 #include <sofa/simulation/Node.h>
 #include <SofaSimulationCommon/SceneLoaderXML.h>
@@ -11,7 +32,7 @@ using sofa::core::execparams::defaultInstance;
 namespace sofa
 {
 
-struct RequiredPlugin_test : public BaseTest
+struct RequiredPlugin_test : public BaseSimulationTest
 {
     void testNotExistingPlugin()
     {
